extracted filter and sort from results to options component
